Move to storer all file from storage process
